+++ CHANGES.md
@@ -31,7 +31,15 @@ To be released.
     its `Equals(object)` method and `GetHashCode()` method became to have
     default behavior of `object` class.  [[#216]]
  -  also, `Swarm` class now does not implement `IDisposable` too. thus
-    `Swarm.Dispose()` was removed too.
+    `Swarm.Dispose()` was removed too. [[#218]]
+ -  `Swarm` became to use queue to broadcast own message. [[#218]]
+     - The methods that broadcast message now aren't async. so they are renamed
+       as below.
+       - `Swarm.BroadcastBlocksAsync()` → `Swarm.BroadcastBlocks()`
+       - `Swarm.BroadcastTxsAsync()` → `Swarm.BroadcastTxs()`
+     - `Swarm` became to receive a `linger` on its constructor. the linger is
+       used to determine how long to wait for a pending message when stopping
+       the swarm.
  -  The type of `Block<T>.Difficulty` is changed to `long` instead of `int`, and
     related classes method parameters and field types have changed accordingly.
  -  Removed `HashDigest.HasLeadingZeroBits()` method.  [[#213]]
@@ -119,6 +127,7 @@ To be released.
 [#215]: https://github.com/planetarium/libplanet/pull/215
 [#216]: https://github.com/planetarium/libplanet/pull/216
 [#217]: https://github.com/planetarium/libplanet/pull/217
+[#218]: https://github.com/planetarium/libplanet/pull/218
 [#223]: https://github.com/planetarium/libplanet/pull/223
 [#231]: https://github.com/planetarium/libplanet/pull/231